requirements.txt
aws-cdk-lib==2.88.0
constructs>=10.0.0,<11.0.0
aws_cdk.aws_glue_alpha

PyYAML~=6.0.3
pyspark~=3.3.0

# Test dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
mock>=4.0.3